pyproject.toml
[project]
name = "botter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "schedule>=1.2.2",
    "secure-smtplib>=0.1.1",
    "types-requests>=2.32.4.20250809",
]
